Bug 1453016 - user safer enum serialization for gfx::SurfaceFormat in IPC; r=kats
MozReview-Commit-ID: Fhyg8e5lyW6

+++ widget/PuppetWidget.cpp
@@ -1037,7 +1037,7 @@ PuppetWidget::SetCursor(imgIContainer* aCursor,
   nsDependentCString cursorData(surfaceData.get(), length);
   mozilla::gfx::IntSize size = dataSurface->GetSize();
   if (!mTabChild->SendSetCustomCursor(cursorData, size.width, size.height, stride,
-                                      static_cast<uint8_t>(dataSurface->GetFormat()),
+                                      dataSurface->GetFormat(),
                                       aHotspotX, aHotspotY, mUpdateCursor)) {
     return NS_ERROR_FAILURE;
   }
